Replace pix array from double to int32_t in render_fast and render_next_pixels_til_tmax_on_back_canvas because it was not useful anymore without anti aliasing, added const keyword where it need to

+++ src/minirt/app/app.h
@@ -15,7 +15,6 @@
 
 # include <stdbool.h>
 # include <stdint.h>
-# include <stdlib.h>
 
 # include "MLX42/MLX42.h"
 
@@ -103,7 +102,7 @@ int32_t			render_ray_from_camera(
 t_ray			create_ray_from_pixel_coords(
 					t_camera const *camera,
 					t_canvas const *canvas,
-					double const coords[2]);
+					int32_t const coords[2]);
 
 /// raytracing
